align test structure to the idiomatic branch
